Fixing / tested Items

+++ Dad-A-Store/Models/Item.cs
@@ -7,7 +7,7 @@ namespace Dad_A_Store.Models
     public Guid   ItemID          { get; set; }
     public string ItemName        { get; set; }
     public string ItemDescription { get; set; }
-    public double ItemPrice       { get; set; }
+    public decimal ItemPrice       { get; set; }
     public Guid   CategoryID      { get; set; }
   }
 }
